+++ src/apps/App.module.css
@@ -1,4 +1,4 @@
-@keyframes huerot {
+@keyframes hue-rotate {
   from {
     filter: hue-rotate(0deg);
   }
@@ -7,11 +7,13 @@
   }
 }
 
-canvas {
+.fit {
   height: 100%;
   width: 100%;
 
   object-fit: cover;
+}
 
-  animation: huerot 7227ms ease-in-out 0s infinite forwards;
+.hue-rotate {
+  animation: hue-rotate 7227ms ease-in-out 0s infinite forwards;
 }
