+++ src/apps/AppTwo.tsx
@@ -1,13 +1,10 @@
-import React, { MouseEventHandler } from 'react'
+import type { Position } from '../types'
 
-import './App.css'
+import React from 'react'
 
-interface Position<T = number> {
-  x: T
-  y: T
-}
+import styles from './App.module.css'
 
-export function AppTwo() {
+export function AppTwo(): React.ReactElement {
   const appString: string = `
     BeelzeAboveBelowOhHowItShowsWhatWeDontKnowYetSeemToBestow. 
     To Say no mo' of what is in store, it's quite a Bore, my lovely deadly whore. 
@@ -54,7 +51,6 @@ export function AppTwo() {
     const canvas: HTMLCanvasElement = ctx.canvas
 
     let frameHandle: number
-    let frameHandleTwo: number
 
     let frameCount: number = 0
 
@@ -146,26 +142,23 @@ export function AppTwo() {
 
     return (): void => {
       window.cancelAnimationFrame(frameHandle)
-      window.cancelAnimationFrame(frameHandleTwo)
     }
   }, [appString, appStringMax])
 
-  const onMouseMove = React.useCallback<MouseEventHandler<HTMLCanvasElement>>(
-    (event: React.MouseEvent<HTMLCanvasElement>): void => {
-      mousePosition.current.x = event.clientX
-      mousePosition.current.y = event.clientY
-    },
-    [],
-  )
+  const onMouseMove = React.useCallback<
+    React.MouseEventHandler<HTMLCanvasElement>
+  >((event: React.MouseEvent<HTMLCanvasElement>): void => {
+    mousePosition.current.x = event.clientX
+    mousePosition.current.y = event.clientY
+  }, [])
 
   return (
-    <>
-      <canvas
-        width={720}
-        height={270}
-        ref={canvasRef}
-        onMouseMove={onMouseMove}
-      />
-    </>
+    <canvas
+      className={`${styles.fit} ${styles['hue-rotate']}`}
+      width={720}
+      height={270}
+      ref={canvasRef}
+      onMouseMove={onMouseMove}
+    />
   )
 }
