+++ src/main.tsx
@@ -1,13 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { Analytics } from '@vercel/analytics/react'
-
 import { createBrowserRouter, redirect, RouterProvider } from 'react-router-dom'
 
-import { Intro } from './Intro.tsx'
-import { App } from './App.tsx'
-import { AppTwo } from './AppTwo.tsx'
-import { AppThree } from './AppThree.tsx'
+import { Analytics } from '@vercel/analytics/react'
+
+import { Intro, App, AppTwo, AppThree } from './apps'
 import { Alic3 } from './Alic3.tsx'
 
 import './index.css'
