+++ src/types/position.ts
@@ -1,0 +1,4 @@
+export interface Position<T = number> {
+  x: T
+  y: T
+}
